+++ packages/shared/src/components/ui/Select/index.stories.tsx
@@ -1,19 +1,4 @@
-import { Select, SelectValue, SelectTrigger, SelectContent, SelectItem } from ".";
-
-const Example = () => {
-  return (
-    <Select>
-      <SelectTrigger className="w-[180px]">
-        <SelectValue placeholder="Theme" />
-      </SelectTrigger>
-      <SelectContent>
-        <SelectItem value="light">Light</SelectItem>
-        <SelectItem value="dark">Dark</SelectItem>
-        <SelectItem value="system">System</SelectItem>
-      </SelectContent>
-    </Select>
-  );
-};
+import { Example } from ".";
 
 import type { Meta, StoryObj } from "@storybook/react";
